requirements.txt
elasticsearch
PyYAML
jinja2

kafka-python
python-snappy
